% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode2.R
\name{fct_recode2}
\alias{fct_recode2}
\title{Manually recode factors with crosstab of old and new}
\usage{
fct_recode2(.f, ..., .verbose = TRUE, .ordered = FALSE,
  .other_to_na = FALSE)
}
\arguments{
\item{.f}{a factor (or character vector)}

\item{...}{named character vectors, name is new level value(s) are old. Levels not mentioned are left as-is unless `other_to_na` is chosen. Levels can be removed by naming NULL.}

\item{.verbose}{print diagnostics? Defaults to TRUE}

\item{.ordered}{order factor according to order supplied? Defaults to FALSE}

\item{.other_to_na}{unmentioned levels to NA? defaults to FALSE}
}
\value{
.f recoded as specified
}
\description{
Sort of like `forcats::fct_recode`, but allows for multiple values per name and adds some extra options.
}
\examples{
fct_recode2(dplyr::starwars$gender, m_or_f = c("male", "female"),
other = c(NA, "none", "hermaphrodite"))
fct_recode2(dplyr::starwars$gender, m_or_f = c("male", "female"),
other = c(NA, "none", "hermaphrodite"), .ordered = TRUE)
fct_recode2(dplyr::starwars$gender, m_or_f = c("male", "female"),
other = c("hermaphrodite"), .other_to_na = TRUE)
}
